Changes made RDCC-3849: Changes made to fix sonar issues
